Servo and speaker/piezo sketch
Using Myo poses to control 3 connected LEDs with pose data printed on an LCD.
Servo is controlled by angles from Quaternions provided by the Myo. 
           
Speaker/piezo output requires q1 to be set (fist) then wave-in to be performed.
This ouput command "tone" is located in HandleIMUData() and will be looped with it.
To stop the speaker/piezo output, perform a different pose (other than wave-in), or reset the lastPose variable by performing the "finger tap" pose.